Extracted identify() function
