+++ src/app/services/website/WebsiteService.live.ts
@@ -1,3 +1,2 @@
 // Use mock implementation until matching against database works
-// tslint:disable no-require-imports
-export const identify = require('./WebsiteService.mock');
+export { default as identify } from './mock/identify';
